Switch return value for the parse_lat() and parse_lon() fuctions from int to double.  We had added some logic to the functions the apply range checks for bad data but failed to update the return value.  Therefore they were getting truncated.
Former-commit-id: 1ec8c5e87aab357fdaa46ab0d2088babe949594f

+++ src/libcode/vx_tc_util/atcf_line_base.cc
@@ -317,8 +317,8 @@ unixtime parse_time(const char *s) {
 
 ////////////////////////////////////////////////////////////////////////
 
-int parse_lat(const char *s) {
-   int v;
+double parse_lat(const char *s) {
+   double v;
 
    v = parse_int(s);
 
@@ -343,8 +343,8 @@ int parse_lat(const char *s) {
 
 ////////////////////////////////////////////////////////////////////////
 
-int parse_lon(const char *s) {
-   int v;
+double parse_lon(const char *s) {
+   double v;
 
    v = parse_int(s);
